fix: CI clippy 零警告（修复 all-targets 下所有 lint）
- unnecessary_sort_by → sort_by_key(Reverse) (session/manager.rs)
- unnecessary_cast i64 → i64 (adapter-telegram)
- 无用 import/convert (adapter-discord, api tests)
- dead_code allow + 冗余借用 (adapter-wechat)
- map_or → is_some_and (integration tests)
- borrow_box lint (plugin-sdk ffi)
- assert_eq!(bool, false) → assert!(!bool) (adapter-qq)

+++ crates/easybot-adapter-telegram/src/lib.rs
@@ -442,7 +442,7 @@ impl PlatformAdapter for TelegramAdapter {
         Ok(SendResult {
             success: true,
             message_id: Some(result.message_id.to_string()),
-            timestamp: Some(result.date as i64 * 1000),
+            timestamp: Some(result.date * 1000),
             error: None,
             error_code: None,
             retryable: false,
@@ -497,7 +497,7 @@ impl PlatformAdapter for TelegramAdapter {
             Ok(SendResult {
                 success: true,
                 message_id: Some(result.message_id.to_string()),
-                timestamp: Some(result.date as i64 * 1000),
+                timestamp: Some(result.date * 1000),
                 error: None,
                 error_code: None,
                 retryable: false,
@@ -549,7 +549,7 @@ impl PlatformAdapter for TelegramAdapter {
                     Ok(SendResult {
                         success: true,
                         message_id: Some(result.message_id.to_string()),
-                        timestamp: Some(result.date as i64 * 1000),
+                        timestamp: Some(result.date * 1000),
                         error: None,
                         error_code: None,
                         retryable: false,
@@ -611,7 +611,7 @@ impl PlatformAdapter for TelegramAdapter {
         Ok(SendResult {
             success: true,
             message_id: Some(result.message_id.to_string()),
-            timestamp: Some(result.date as i64 * 1000),
+            timestamp: Some(result.date * 1000),
             error: None,
             error_code: None,
             retryable: false,
